backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contract-review-backend"
version = "0.1.0"
description = "FastAPI backend for the Contract Review System."
authors = [
  { name = "AI Contract Analyzer" }
]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "SQLAlchemy>=2.0",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "celery>=5.3",
  "redis>=5.0",
  "python-multipart>=0.0.9",
  "pypdf>=4.0.0",
  "httpx>=0.27.0",
  "requests>=2.31.0",
  "tenacity>=8.2.0",
  "python-docx>=1.1.0",
  "docx2python>=2.0.0",  # Preserves numbered lists in DOCX (python-docx loses them)
  "pytesseract>=0.3.10",
  "Pillow>=10.0.0",
  "pdf2image>=1.17.0",
  # Advanced parsing dependencies
  "pymupdf>=1.23.0",  # Primary PDF parser (fast and accurate)
  "pdfplumber>=0.11.0",  # Table extraction (primary method)
  "camelot-py[cv]>=0.11.0",  # Advanced table extraction for complex tables
  "pandas>=2.0.0",
  "openpyxl>=3.1.0",
  "python-pptx>=0.6.0",
  "beautifulsoup4>=4.12.0",
  "html5lib>=1.1",
  "lxml>=5.0.0",
  "markdown>=3.5.0",
  "chardet>=5.2.0",
  "xgboost>=2.0.0",
  "scikit-learn>=1.4.0",
  "numpy>=1.26.0",
  # ML dependencies removed - using regex-based extraction instead
  # "torch>=2.0.0",  # Removed: 800MB, not needed
  # "torchvision>=0.15.0",  # Removed: 300MB, not needed
  # "transformers>=4.30.0",  # Removed: 400MB, DocFormer not used
  # "onnxruntime>=1.16.0",
  # "huggingface-hub>=0.20.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "httpx>=0.26",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
  "black>=24.0",
  "mypy>=1.8"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]
target-version = "py311"

[tool.ruff.isort]
known-first-party = ["app"]
combine-as-imports = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

